modificaciones de la tarea 8

+++ Data/Models/Vacunado.cs
@@ -25,8 +25,8 @@ namespace Tarea7.Data.Models
         public DateTime? Vacuna1Fecha { get; set; }
         public int? Vacuna2Id { get; set; }
         public DateTime? Vacuna2Fecha { get; set; }
-        public double? Latitud { get; set; }
-        public double? Longitud { get; set; }
+        public float? Latitud { get; set; }
+        public float? Longitud { get; set; }
 
         public virtual Provincia Provincia { get; set; }
         public virtual Vacuna Vacuna1 { get; set; }
